+++ src/org/commoncrawl/tutorial/WordCountReducer.java
@@ -2,7 +2,6 @@ package org.commoncrawl.tutorial;
 
 import java.io.IOException;
 import java.util.Iterator;
-import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapred.MapReduceBase;
@@ -11,8 +10,8 @@ import org.apache.hadoop.mapred.Reducer;
 import org.apache.hadoop.mapred.Reporter;
 
 public class WordCountReducer extends MapReduceBase
-  implements Reducer<Text, IntWritable, Text, LongWritable> {
-  public void reduce(Text key, Iterator<IntWritable> values,
+  implements Reducer<Text, LongWritable, Text, LongWritable> {
+  public void reduce(Text key, Iterator<LongWritable> values,
 	      OutputCollector<Text, LongWritable> output, Reporter reporter)
 	      throws IOException {
 	long sum = 0;
